+++ detectors/aws/ecs/ecs_test.go
@@ -16,7 +16,10 @@ package ecs
 
 import (
 	"context"
+	http "net/http"
+	"net/http/httptest"
 	"os"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -42,11 +45,10 @@ func (detectorUtils *MockDetectorUtils) getContainerName() (string, error) {
 	return args.String(0), args.Error(1)
 }
 
-// successfully return resource when process is running on Amazon ECS environment.
-func TestDetect(t *testing.T) {
+// succesfully return resource when process is running on Amazon ECS environment with no Metadata v4.
+func TestDetectV3(t *testing.T) {
 	os.Clearenv()
 	_ = os.Setenv(metadataV3EnvVar, "3")
-	_ = os.Setenv(metadataV4EnvVar, "4")
 
 	detectorUtils := new(MockDetectorUtils)
 
@@ -63,7 +65,52 @@ func TestDetect(t *testing.T) {
 	detector := &resourceDetector{utils: detectorUtils}
 	res, _ := detector.Detect(context.Background())
 
-	assert.Equal(t, res, expectedResource, "Resource returned is incorrect")
+	assert.Equal(t, expectedResource, res, "Resource returned is incorrect")
+}
+
+//succesfully return resource when process is running on Amazon ECS environment with Metadata v4.
+func TestDetectV4(t *testing.T) {
+	testServer := httptest.NewServer(http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
+		if strings.HasSuffix(req.URL.String(), "/task") {
+			content, err := os.ReadFile("testdata/metadatav4-response-task.json")
+			if err == nil {
+				res.Write(content)
+			}
+		} else {
+			content, err := os.ReadFile("testdata/metadatav4-response-container.json")
+			if err == nil {
+				res.Write(content)
+			}
+		}
+	}))
+	defer func() { testServer.Close() }()
+
+	os.Clearenv()
+	_ = os.Setenv(metadataV3EnvVar, "3")
+	_ = os.Setenv(metadataV4EnvVar, testServer.URL)
+
+	detectorUtils := new(MockDetectorUtils)
+
+	detectorUtils.On("getContainerName").Return("container-Name", nil)
+	detectorUtils.On("getContainerID").Return("0123456789A", nil)
+
+	attributes := []attribute.KeyValue{
+		semconv.CloudProviderAWS,
+		semconv.CloudPlatformAWSECS,
+		semconv.ContainerNameKey.String("container-Name"),
+		semconv.ContainerIDKey.String("0123456789A"),
+		semconv.AWSECSContainerARNKey.String("arn:aws:ecs:us-west-2:111122223333:container/0206b271-b33f-47ab-86c6-a0ba208a70a9"),
+		semconv.AWSECSClusterARNKey.String("arn:aws:ecs:us-west-2:111122223333:cluster/default"),
+		semconv.AWSECSLaunchtypeKey.String("EC2"),
+		semconv.AWSECSTaskARNKey.String("arn:aws:ecs:us-west-2:111122223333:task/default/158d1c8083dd49d6b527399fd6414f5c"),
+		semconv.AWSECSTaskFamilyKey.String("curltest"),
+		semconv.AWSECSTaskRevisionKey.String("26"),
+	}
+	expectedResource := resource.NewWithAttributes(semconv.SchemaURL, attributes...)
+	detector := &resourceDetector{utils: detectorUtils}
+	res, _ := detector.Detect(context.Background())
+
+	assert.Equal(t, expectedResource, res, "Resource returned is incorrect")
 }
 
 // returns empty resource when detector cannot read container ID.